A workaround for a race condition with zsh.
zsh exhibits a race condition upon receiving a SIGCHLD. Since it
happens when calling free() inside and outside of a signal handler at
the same time, I'm not sure whether it's zsh or glibc to blame. The
workaround is simple: disown the job immediately. Luckily, zsh allows
to do it immediately using &!, otherwise, there would not be much
point.

bash and dash don't have these problems.