Make it explicit that test header is sourced from current directory.
It turns out that sourcing a file without an explicit path makes shell
search PATH. The exception is bash when not run as sh. Funny enough, all
tested shells behave just like bash when 'make test' is run from command
line. It was only during build-time testing on gentoo that this bug was
discovered: in an ebuild, shells behave as they should. Go figure...